==> notebooks/base-strategy/v1/2019-05-09 supports classification/07-strategy-with-base-model.ipynb <==
%run ../../../utils/commonImports.py
%run ../../../utils/tradingImports.py
%matplotlib inline

from itertools import product
from base import *
from exputils import *

# Load data

# dropbox_dir = 'D:\\Dropbox\\My work\\krypl-project'
dropbox_dir = '/Users/david.vesely/Dropbox/My work/krypl-project'
features = ['change_1', 'change_2', 'change_3']
pairs = read_tsv('selected-pairs.tsv', header=None)[0].tolist()

train_dict = {p: read_train(p, dropbox_dir) for p in pairs}
ohlc_featured_train = load_model('ohlc_featured_train.pkl')

model = load_model('models/lin_regression.pkl')

# Setup

features = ['support', 'change_1', 'change_2', 'change_3']

window_sizes = [4]
shifts = [1, 2, 3]

DAY = 48
max_hold_time = 5 * DAY
stop_loss = 0.05
fee = 0.0025

min_profit_targets = [round(i / 100, 2) for i in range(0, 6)]

args = product(pairs, window_sizes, min_profit_targets)
args = list(args)
len(args)

# Train trading stats

def trading_stats(data, pair, willing_loss, min_profit_target):
    contract_pair = ContractPair.new(*[x.lower() for x in pair.split('_')])
    dm = OhlcDataManager(data[['open', 'low', 'high', 'close']], data[features])
    wallet = {'btc': 100}
    exchange = BackTestExchange(dm, wallet, 0.0025)
    strategy = BaseMLStrategy(exchange, 
                            dm, 
                            contract_pair, 
                            willing_loss=willing_loss,  
                            min_profit_target=min_profit_target,
                            model=model)
    strategy.trade()
    strategy_stats = strategy.stats(contract_pair['priceContract']).report()
    return strategy_stats


def trading_stats_f(args):
    pair, window_size, min_profit_target = args
    
    data = train_dict[pair]
    data[features] = ohlc_featured_train[pair][window_size]
    stats = trading_stats(data, pair, stop_loss, min_profit_target)
    return pair, window_size, stop_loss, min_profit_target, stats

results = run_parallel(trading_stats_f, args)

param_cols = ['pair', 'window_size', 'stop_loss', 'min_profit_target']
stats_cols = ['startAmount', 'numberOfTrades', 'totalProfit', 'avgProfit',
    'winPercentage', 'avgWinTrade', 'avgLossTrade', 'profitFactor',
    'maxDrawdown', 'avgTimeToClose', 'minTimeToClose', 'maxTimeToClose']

result_df = pd.DataFrame(results, columns=param_cols+['stats'])

for c in stats_cols:
    result_df[c] = result_df['stats'].apply(lambda x: x.iloc[:, 0][c])
    
result_df = result_df.drop('stats', axis=1)
result_df.to_pickle('train_results.pkl')

result_df

# Tmp

from modeling.strategy import Strategy


def trading_stats2(data, pair, willing_loss, min_profit_target):
    contract_pair = ContractPair.new(*[x.lower() for x in pair.split('_')])
    dm = OhlcDataManager(data[['open', 'low', 'high', 'close']], data[features])
    wallet = {'btc': 100}
    exchange = BackTestExchange(dm, wallet, 0.0025)
    strategy = BaseMLStrategy(exchange, 
                            dm, 
                            contract_pair, 
                            willing_loss=willing_loss,  
                            min_profit_target=min_profit_target,
                            model=model)
    strategy.trade()
    strategy_stats = strategy.stats(contract_pair['priceContract'])
    return strategy_stats

pair = 'BTC_ETH'
data = train_dict[pair]
data[features] = ohlc_featured_train[pair][4]
min_profit_target = 0.03

stats.report()
