==> notebooks/base-strategy/v1/2019-05-09 supports classification/08-select-strategy.ipynb <==
%run ../../../utils/commonImports.py
%run ../../../utils/tradingImports.py
%matplotlib inline

from itertools import product
from base import *
from exputils import *

# Load data

# dropbox_dir = 'D:\\Dropbox\\My work\\krypl-project'
dropbox_dir = '/Users/david.vesely/Dropbox/My work/krypl-project'
features = ['change_1', 'change_2', 'change_3']
pairs = read_tsv('selected-pairs.tsv', header=None)[0].tolist()

val_dict = {p: read_val(p, dropbox_dir) for p in pairs}
ohlc_featured_train = load_model('ohlc_featured_val.pkl')

model = load_model('models/lin_regression.pkl')

results = pd.read_pickle('train_results.pkl')

# Setup

features = ['support', 'change_1', 'change_2', 'change_3']

window_sizes = [4]
shifts = [1, 2, 3]

DAY = 48
max_hold_time = 5 * DAY
stop_loss = 0.05
fee = 0.0025

min_profit_targets = [round(i / 100, 2) for i in range(0, 6)]

args = product(pairs, window_sizes, min_profit_targets)
args = list(args)
len(args)

# Select params

def min_max(x):
    return (x - x.min()) / (x.max() - x.min())

results['zTotalProfit'] = min_max(results['totalProfit'])
results['zWinPercentage'] = min_max(results['winPercentage'])
results['zAvgLossTrade'] = min_max(results['avgLossTrade'])
results['meanStats'] = (results['zTotalProfit'] + 
                        (2 * results['zWinPercentage']) + 
                        results['zAvgLossTrade'] * (1 - results['winPercentage'] / 100)) / 3

selected_params = pd.DataFrame()
for pair in pairs:
    p_results = results[results['pair'] == pair]
    p_results = reset_index_hard(p_results).copy()
    row = p_results.query('totalProfit > 0')\
        .query('-avgLossTrade <= avgWinTrade')\
        .sort_values('totalProfit', ascending=False)
    if len(row) > 0:
        selected_params = selected_params.append(row.iloc[0, :])

cols = ['pair', 'window_size', 'stop_loss', 'min_profit_target', 
         'totalProfit', 'maxTimeToClose', 'winPercentage', 'avgLossTrade', 'avgWinTrade']
selected_params[cols]
