==> notebooks/base-strategy/v1/2019-05-09 supports classification/09-conclusion.ipynb <==
# Conclusion

Linearni regrese sve targety spise prestreluje a v zavislosti na tom, je mensi winPercentage.

V dalsim experimentu by to chtelo zkusit tento problem vyresit napriklad jinou ucici funkci nebo jinym labelovanim..

Slo by napriklad olabelovat data treba aby target byl 90% max profitu
